Improve navigation performance (#42)
* Do not call setView all the time

* Remove debug console count

* Defer focus by some ms

This helps both AT announcements and avoids layout trashing

* Scroll to top when routing, combine with focus

And some fun ordering of macro tasks and animation frames

* Debounce setInitialViewOnce

I think this is the best we could do with setting the features
automatically. I really like it; it gives a consistent view on
load, and does not bother past a deadline. Works for size
calculations as well.

* Remove window.innerWidth|Height from system-info

They would force/invalidate the layout. I don't think we need
these anyway.

* Prevent small flash/re-layout in attribution

This happened because of the deferred setView. We now include
the attribution as a prefix, keeping it always visible, from
the get-go.

* Simplify flex propagation, use vh for leaflet-map

* Clean up some comments, css properties

* Try .sleep 200 vs 100

* Revert "Try .sleep 200 vs 100"

This reverts commit 5b1a7b33428c82caa7a13e39efb8882bb4e63df5.